+++ recipes/attribute_driven_domain.rb
@@ -93,6 +93,13 @@ Another approach using a vagrant file is to set the json attribute such as;
                         'securityenabled' => false
                       }
                     },
+                    'network_listeners' => {
+                      'jk-listener' => {
+                        'listenerport' => 8009,
+                        'jkenabled' => true,
+                        'protocol' => 'http-listener-1'
+                      }
+                    },
                     'context_services' => {
                       'concurrent/MyAppContextService' => {
                         'description' => 'My Apps ContextService'
@@ -462,6 +469,28 @@ gf_sort(node['glassfish']['domains']).each_pair do |domain_key, definition|
       securityenabled config['securityenabled'] unless config['securityenabled'].nil?
       maxqueuesize config['maxqueuesize'] if config['maxqueuesize']
       properties config['properties'] if config['properties']
+    end
+  end
+ 
+  Chef::Log.info "Defining GlassFish Domain #{domain_key} - network listeners"
+  gf_sort(definition['network_listeners'] || {}).each_pair do |key, config|
+    glassfish_network_listener key do
+      domain_name domain_key
+      admin_port admin_port if admin_port
+      username username if username
+      password_file password_file if password_file
+      secure secure if secure
+      system_user system_username if system_username
+      system_group system_group if system_group
+
+      address config['address'] if config['address']
+      listenerport config['listenerport'] if config['listenerport']
+      threadpool config['threadpool'] if config['threadpool']
+      protocol config['protocol'] if config['protocol']
+      transport config['transport'] if config['transport']
+      enabled config['enabled'] if config['enabled']
+      jkenabled config['jkenabled'] if config['jkenabled']
+      target config['target'] if config['target']
     end
   end
 
@@ -1262,6 +1291,24 @@ gf_sort(node['glassfish']['domains']).each_pair do |domain_key, definition|
     unless definition['iiop_listeners'] && definition['iiop_listeners'][existing]
       Chef::Log.info "Defining GlassFish Domain #{domain_key} - removing existing iiop-listener #{existing}"
       glassfish_iiop_listener existing do
+        domain_name domain_key
+        admin_port admin_port if admin_port
+        username username if username
+        password_file password_file if password_file
+        secure secure if secure
+        system_user system_username if system_username
+        system_group system_group if system_group
+        action :delete
+      end
+    end
+  end
+
+  Chef::Log.info "Defining GlassFish Domain #{domain_key} - checking existing network_listeners"
+  gf_scan_existing_resources(admin_port, username, password_file, secure, 'list-network-listeners') do |existing|
+    Chef::Log.info "Defining GlassFish Domain #{domain_key} - considering existing network_listeners #{existing}"
+    unless definition['network_listeners'] && definition['network_listeners'][existing]
+      Chef::Log.info "Defining GlassFish Domain #{domain_key} - removing existing network-listener #{existing}"
+      glassfish_network_listener existing do
         domain_name domain_key
         admin_port admin_port if admin_port
         username username if username